requirements.txt
fastapi
uvicorn
dnspython
aiosmtplib
python-multipart
sqlalchemy
slowapi
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
celery
redis
psycopg2-binary
gunicorn
sendgrid

aioredis
stripe